feat: get isLogin status from redux
